Fix week number bug.
This is related to that "from the future" comment I made previously.

Neither the observation packets nor the ION / UTC data packets contain a
week number or anything that could be turned into a week number. Permit
the starting week number to be specified with a receiver option (WEEK=n).
Default to the current week if not specified. Handle week rollover and
increment of the specified week number.