Update ublox.c
The FS32 type is not working as expected as the "argument" string is not null terminated (The ublox reference manual describes this type as "32 ASCII bytes including 0-termination").
Further using a string length of 32 results in the string terminator of the output string overwritten by the next argument. The changed code will leave the remaining part of the 31 char string uninitialized which should not be a problem.