Fix mistakes
Fix mistakes. Note none of this will actually run because the
infrastructure to make it work does not appear to be in place.
Specifically raw->strfmt is not yet set by anything in RTKLIB, etc. What
would set it and where I do not know. In the mean time I'm in limbo.