Fix potential deadlock when starting / stopping rtk sever quickly
The rtksvr thread uses svr->state to break out of the sever loop
and exit the thread. Thread termination is done b setting svr->state
in rtksvrstop. As the thread is responsible for setting svr->state
to 1 there is a race condition here. Instead set the state variable
in rtksvrstart if thread creation is successful.